Basic detection of blank areas
All boxes that form a blank area (area with no number, which means no
bomb touching the box) share a same link target, so that they will
get "opened" at the same time when one of them is clicked.

This still differs from standard minesweeper though, in that the
numbers that surround the blank area don't get unveiled.