Use sprite for the smiley button
In Win95 style, use a single sprite image for the "smiley button"
instead of 3 separate images for new, fear and lose states. This
allows to avoid lag when performing an action (e.g. losing) for the
1st time.

Also draw the border around the button via CSS instead of integrating
the border inside the image. This allows to simulate a nice "push"
effect when the button is pressed.